+++ quasar.conf.js
@@ -75,7 +75,8 @@ module.exports = function (ctx) {
         'QList',
         'QListHeader',
         'QTooltip',
-        'QPopover'
+        'QPopover',
+        'QChip'
       ],
       plugins: [
         'Loading',
@@ -89,7 +90,6 @@ module.exports = function (ctx) {
     animations: [
     ],
     pwa: {
-      cacheExt: 'js,html,css,ttf,eot,otf,woff,woff2,json,svg,gif,jpg,jpeg,png,wav,ogg,webm,flac,aac,mp4,mp3',
       manifest: {
         name: 'TelemetryViewer',
         short_name: 'Telemetry-Viewer',