+++ src/store/mutations.js
@@ -1,5 +1,5 @@
 import Vue from 'vue'
-import { Cookies, LocalStorage } from 'quasar'
+import { Cookies, LocalStorage, Notify } from 'quasar'
 
 function reqStart (state) {
   if (DEV) {
@@ -16,28 +16,49 @@ function reqFailed (state, payload) {
     console.log('Failed Request')
     console.log(payload)
   }
-  console.log(payload)
-  switch (payload.response.status) {
-    case 0: {
-      setOfflineFlag(state, true)
-      Vue.set(state, 'token', '')
-      break
-    }
-    case 401: {
-      clearToken(state)
-      break
-    }
-    default: {
-      if (DEV) {
-        console.log(`${payload.status} - ${payload.statusText}`)
+  /* http errors */
+  if ((payload.response && payload.response.status)) {
+    switch (payload.response.status) {
+      case 0: {
+        setOfflineFlag(state, true)
+        unsetDevicesInit(state)
+        Vue.set(state, 'token', '')
+        break
+      }
+      case 401: {
+        clearToken(state)
+        break
+      }
+      default: {
+        if (DEV) {
+          console.log(`${payload.status} - ${payload.statusText}`)
+        }
+        if (payload.response.data && payload.response.data.errors && payload.response.data.errors.length) {
+          payload.response.data.errors.forEach((e) => { addError(state, e.reason) })
+        }
       }
     }
+    /* mqtt errors */
+  } else if (payload.code && payload.message) {
+    switch (payload.code) {
+      case 2: {
+        if (state.token) {
+          clearToken(state)
+        }
+        addError(state, payload.message)
+        break
+      }
+    }
+  } else {
+    addError(state, payload.message)
   }
 }
 function setToken (state, val) {
   let token = val.replace('FlespiToken ', '')
+  setSocketOffline(state, true)
   if (val && token.match(/^[a-z0-9]+$/i)) {
     Vue.connector.token = `FlespiToken ${token}`
+    Vue.connector.socket.on('connect', () => { setSocketOffline(state, false) })
     LocalStorage.set('X-Flespi-Token', token)
   } else {
     token = ''
@@ -45,6 +66,10 @@ function setToken (state, val) {
     clearToken(state)
   }
   Vue.set(state, 'token', token)
+  if (state.errors.length) {
+    state.errors = []
+    state.newNotificationCounter = 0
+  }
 }
 function clearToken (state) {
   const cookieToken = Cookies.get('X-Flespi-Token'),
@@ -54,6 +79,7 @@ function clearToken (state) {
   }
   LocalStorage.remove('X-Flespi-Token')
   Vue.connector.token = ''
+  if (state.socketOffline) { setSocketOffline(state, false) }
   Vue.set(state, 'token', '')
 }
 function setDevicesInit (state) {
@@ -67,6 +93,54 @@ function setOfflineFlag (state, flag) {
   Vue.set(state, 'offline', flag)
 }
 
+function addError (state, message) {
+  Notify.create({
+    type: 'negative',
+    icon: 'warning',
+    message: `${message}`,
+    timeout: 1000
+  })
+  state.newNotificationCounter++
+  state.errors.push(message)
+}
+
+function clearErrors (state) {
+  Vue.set(state, 'errors', [])
+}
+
+function setSocketOffline (state, flag) {
+  Vue.set(state, 'socketOffline', flag)
+}
+
+function updateDevices (state, payload) {
+  switch (payload.type) {
+    case 'created': {
+      state.devices[payload.device.id] = payload.device
+      break
+    }
+    case 'updated': {
+      Object.keys(state.devices).some((id) => {
+        if (id === payload.device.id) {
+          state.devices[id] = Object.assign(state.devices[id], payload.device)
+          return true
+        }
+        return false
+      })
+      break
+    }
+    case 'deleted': {
+      Object.keys(state.devices).some((id, index) => {
+        if (id === payload.device.id) {
+          delete state.devices[id]
+          return true
+        }
+        return false
+      })
+      break
+    }
+  }
+}
+function clearNotificationCounter (state) { state.newNotificationCounter = 0 }
 export default {
   reqStart,
   reqSuccessful,
@@ -75,5 +149,10 @@ export default {
   clearToken,
   setDevicesInit,
   unsetDevicesInit,
-  setOfflineFlag
+  setOfflineFlag,
+  addError,
+  clearErrors,
+  setSocketOffline,
+  updateDevices,
+  clearNotificationCounter
 }